pyproject.toml
[tool.poetry]
name = "ghostpost-be"
version = "0.1.0"
description = "Backend API for GhostPost"
authors = ["Kathryn Anderson <kat42272@yahoo.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.12.1"
django-cors-headers = "^3.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
